+++ inst/include/tools/SlicingIndex.h
@@ -4,9 +4,10 @@
 class SlicingIndex {
 public:
     
-    SlicingIndex(IntegerVector data_) : data(data_){}
+    SlicingIndex(IntegerVector data_) : data(data_), group_index(-1) {}
+    SlicingIndex(IntegerVector data_, int group_) : data(data_), group_index(group_) {}
     
-    SlicingIndex(int start, int n) : data(seq(start,start+n-1)){}
+    SlicingIndex(int start, int n) : data(seq(start,start+n-1)), group_index(-1) {}
     
     inline int size() const { 
         return data.size() ; 
@@ -16,8 +17,11 @@ public:
         return data[i] ;    
     }
     
+    inline int group() const { return group_index ; }
+    
 // private:
     IntegerVector data ;
+    int group_index ;
 } ;
 
 #endif